Fixes for IC-7610 Extra Controls - Band Tab
  * IC-7610 Extra Controls - Band Tab -
    Fixes for band selection and register store functions
  * Add documentation of this feature

+++ src/include/icom/ICbase.h
@@ -46,7 +46,7 @@ public:
 	virtual ~RIG_ICOM() {}
 	void checkresponse();
 	bool sendICcommand(std::string str, int nbr);
-	bool  waitFB(const char *sz, int timeout = 500);
+	bool  waitFB(const char *sz, unsigned long timeout = 500);
 	bool  waitFOR(size_t n, const char *sz, unsigned long timeout = 500);
 	void adjustCIV(uchar adr);
